pyproject.toml
[project]
name = "wttrinbot"
version = "0.0.0a"
description = "An inline Telegram bot to see the current weather"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncpg>=0.30.0",
    "certifi>=2025.8.3",
    "geopy>=2.4.1",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.6",
    "sqlalchemy>=2.0.42",
]

[dependency-groups]
dev = [
    "black>=24.8.0",
    "isort>=6.0.1",
]

[tool.isort]
profile = "black"
